fix some codes

+++ frontend/src/components/WorkoutDetails.js
@@ -9,11 +9,15 @@ const WorkoutDetails = ({ workout }) => {
   const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if(!user) {
+      return
+    }
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: 'DELETE',
       headers: { 
         'Content-Type': 'application/json', 
-        'authorization': user.token 
+        'Authorization': `Bearer ${user.token}`
       }
     })
     const json = await response.json()
